spmd: Fix signedness comparison warning
With -Wsign-compare, compilers issue a warning in the SPMD code:
====================
services/std_svc/spmd/spmd_pm.c:35:22: error: comparison of integer
expressions of different signedness: 'int' and 'unsigned int'
[-Werror=sign-compare]
   35 |  if ((id < 0) || (id >= PLATFORM_CORE_COUNT)) {
      |                      ^~
cc1: all warnings being treated as errors
====================

Since we just established that "id" is positive, we can safely cast it
to an unsigned type to make the comparison have matching types.

+++ services/std_svc/spmd/spmd_pm.c
@@ -32,7 +32,7 @@ int spmd_pm_secondary_core_set_ep(unsigned long long mpidr,
 {
 	int id = plat_core_pos_by_mpidr(mpidr);
 
-	if ((id < 0) || (id >= PLATFORM_CORE_COUNT)) {
+	if ((id < 0) || ((unsigned int)id >= PLATFORM_CORE_COUNT)) {
 		ERROR("%s inconsistent MPIDR (%llx)\n", __func__, mpidr);
 		return -EINVAL;
 	}
